Moved ProjectInfoVerifier with dependencies to the project package
