+++ webhook-service/test/ch/datascience/webhookservice/hookcreation/HookCreationGenerators.scala
@@ -18,20 +18,14 @@
 
 package ch.datascience.webhookservice.hookcreation
 
-import ch.datascience.generators.Generators.validatedUrls
-import ch.datascience.webhookservice.hookcreation.ProjectHookCreator.ProjectHook
-import ch.datascience.webhookservice.hookcreation.ProjectHookUrlFinder.ProjectHookUrl
-import org.scalacheck.Gen
 import ch.datascience.graph.events.EventsGenerators._
 import ch.datascience.webhookservice.generators.ServiceTypesGenerators._
 import ch.datascience.webhookservice.hookcreation.LatestPushEventFetcher.PushEventInfo
-import ch.datascience.webhookservice.hookcreation.ProjectHookVerifier.HookIdentifier
+import ch.datascience.webhookservice.hookcreation.ProjectHookCreator.ProjectHook
 import ch.datascience.webhookservice.hookcreation.UserInfoFinder.UserInfo
+import org.scalacheck.Gen
 
 private object HookCreationGenerators {
-
-  implicit val projectHookUrls: Gen[ProjectHookUrl] =
-    validatedUrls map (url => ProjectHookUrl.apply(url.value))
 
   implicit val projectHooks: Gen[ProjectHook] = for {
     projectId           <- projectIds
@@ -43,16 +37,6 @@ private object HookCreationGenerators {
       hookUrl,
       serializedHookToken
     )
-
-  implicit val projectHookIds: Gen[HookIdentifier] = for {
-    projectId <- projectIds
-    hookUrl   <- projectHookUrls
-  } yield
-    HookIdentifier(
-      projectId,
-      hookUrl,
-    )
-
   implicit val pushEventInfos: Gen[PushEventInfo] = for {
     projectId <- projectIds
     userId    <- userIds
